Bug fix: Avoiding error when cellPopulation = 'All'
Current code tested whether each cell population was present in tile results, and threw an error if it wasn't present. Which meant, if you set cellPopulations to 'All', then  an error was thrown. It also tested each population at a time. 

New code fixes the error, and replaces a for loop with the 'all()' command for greater efficiency.