Updates for SummarizedExperiment output
Updated getSampleTileMatrix and related functions so that we output a summarized experiment that is a matrix of intensities across all samples/cell types bundled into one object. Assays are cell types.

The rowRanges output will have a columns of TRUE and FALSE for whether a given column was actually called within a given cell type.