Fixing no cell error
When a sample has no cells of a given cell type, then the fragslist is null and the sample is tossed out. This creates an issue downstream when the sample tile matrix is generated. This code returns a NULL when the frags for a given celltype-sample is NULL, thus preserving empty behavior.