Improving memory usage
Rather than using pblapply with cores set to a number (which wraps mclapply), I initialize multiple cores, and pass it the variables it needs, thus keeping memory down. When pblapply is called, it's handed those cores, which it passses off to parlapply. The end result is much, much less memory used.