Update for FIgure 4
Adding ability to deal with mismatched regions to tiles. If the user provides a region that doesn't match an exact time, then the function will throw an error unless approximateTile is set to TRUE.

If it is, then the function will match up all tiles that overlap with the regions provided.