Adding test for input object
The function will now test whether the GRanges object has a column with the names FDR and Log2FC_C for testing, and it will test after it identifies all the TSSs. We want to do this because someone may want to simply identify all the TSSs within a given tile set, without running differentials. If they set returnAllTSS, then it will return the TSS set without hitting the new test. If they don't set returnAllTSS to TRUE, then it will look for the FDR and Log2FC_C columns and hit errors, so the test will catch that before it hits those errors and return a message that is actually helpful.