Implementing corrections for memory leaks
Replaced mclapply and parlapply statements with a simple iteration over a list where each index is a sublist that contains all variables for the function.
Instead of initializing a new function within the *lapply statement, I include a function that takes one variable as a list, instead of a series of separate variables.