BUG: Metadata missing & sample names
Needed to reorganize portions of these functions so that it would only test peaks that were called for that cell type.

That involved subsetting down to the cell type, and then running combineSampleTileMatrix that way Fragment and CellNumber metadata became apart of the colData output, allowing us to include them in the model.