ENH: reorganizes output for efficiency
Now outputs a SummarizedExperiment with coefficients, residuals, and variance.

Each assay is a coefficient. Residuals and variance are stored in metadata.

Also adds a function for efficiently processing model outputs, which can be used for other modeling approaches as well (it's used for runZIGLMM too).

Also added edits for pilotLMEM.