Update getPopFrags.R
Changing error messages to be more informative. 
Instead of just returning an error message about cellSubsets, it now lists cellPopulations as well. 
cellPopulations is a variable used in callOpenTiles, which calls getCellPopFrags. However, getCellPopFrags uses cellSubsets. Hence the error messages were useful for getCellPopFrags, but not callOpenTiles. This error fixes it.